% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_bb_metadata}
\alias{get_bb_metadata}
\title{metadata query submitter}
\usage{
get_bb_metadata(query_type = "cell_type", query_val = "GM12878")
}
\arguments{
\item{query_type}{character(1) defaults to "cell_type"}

\item{query_val}{character(1) defaults to "GM12878"}
}
\description{
metadata query submitter
}
\examples{
q = bedbaseRClient:::metadata_query_builder()
q
if (interactive()) {
  md = get_bb_metadata()
  md # use httr::content,...
}
}
