% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_2021_03_09}
\alias{meta_2021_03_09}
\title{metadata table for encode components of bedbase}
\format{
data.frame
}
\usage{
meta_2021_03_09
}
\description{
metadata table for encode components of bedbase
}
\note{
A column "cell_type_onto" was added, after searching UBERON, Cell Ontology, Cell Line Ontology
for exact matches (module 'cell' in Cell Line Ontology), using ontoProc images in Bioc 3.12.
}
\examples{
names(meta_2021_03_09)
head(meta_2021_03_09)
}
\keyword{datasets}
