% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{metadata_query_builder}
\alias{metadata_query_builder}
\title{very speculative approach to creating a query}
\usage{
metadata_query_builder(token1 = "cell_type", query_val = "GM12878")
}
\arguments{
\item{token1}{character(1) defaults to "cell_type"}

\item{query_val}{character(1) defaults to "GM12878" could be any supported cell type or line tag}
}
\value{
character()
}
\description{
very speculative approach to creating a query
}
