% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_bb}
\alias{query_bb}
\title{query a bedbase bigbed file}
\usage{
query_bb(md5sum, ...)
}
\arguments{
\item{md5sum}{character(1) md5sum used as key into bedbase resources}

\item{\dots}{passed to rtracklayer::import.bb, which will have format="bigBed" set}
}
\value{
GRanges
}
\description{
query a bedbase bigbed file
}
\examples{
b1 = query_bb(md5sum="78c0e4753d04b238fc07e4ebe5a02984", 
    which=GenomicRanges::GRanges("chr17:37000000-39000000"))
b1
}
