% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{metadata_query_builder}
\alias{metadata_query_builder}
\title{very speculative approach to creating a query}
\usage{
metadata_query_builder(token1 = "cell_type", query_val = "GM12878")
}
\description{
very speculative approach to creating a query
}
